requirements.txt
apscheduler
boto3
hikari
hikari-lightbulb
pandas
pillow
psycopg